% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Amargo.R
\name{Z}
\alias{Z}
\title{Calculo de proporciones de dos grupos mediante Distribucion Z}
\usage{
Z(x1, n1, x2, n2)
}
\arguments{
\item{x1}{numero de acertados en un grupo 1}

\item{n1}{numero total de la poblacion de muestra 1}

\item{x2}{numero de acertados en un grupo 2}

\item{n2}{numero total de la poblacion de muestra 2}
}
\value{
El resultado a comparar con las tablas Z y comprobar si infiere o no, desarrollando los parametros de cada grupo dentro de la función.
}
\description{
#Realiza el calculo de las proporciones de
dos grupos independientes
}
